Change hadoop workflow key to bypass mapreduce workflow key (#271)
We noticed a problem though where the Pig and Hive frameworks would override 'mapreduce.workflow.id' with its own Pig/Hive workflow ID name. So this PR is to change the configuration key which we inject the azkaban project/flow name to prevent the project/flow name from being overwritten.