Introduce HTML viewer to display html file on azkaban (#280)
Currently html files are treated as plain text on azkaban, for super simple html files, that works fine since they are still readable. For complicated and large html files, this will be unreadable. The new html viewer allows user to open up html files and view it like an actual html file. This viewer is used to display files with ".htm" or ".html" extension.