Introduce HTML viewer to display HTML file on Azkaban server (#284)
Currently html files are treated as text files on Azkaban server, its source code is displayed, which for most of the html files, is too complicated to work with. There are also needs for people to see the rendered html page instead of source code. This patch introduces a html viewer and modifies the template so that html file content is rendered. To protect people from malicious html files, the content is rendered within a sandboxed iframe, which prevents any javascript execution.